src/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
async-timeout==5.0.1
backports.zoneinfo==0.2.1
billiard==4.2.1
blinker==1.8.2
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.1.1
cycler==0.12.1
flask==3.0.3
fonttools==4.55.0
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
itsdangerous==2.2.0
jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
MarkupSafe==2.1.5
matplotlib==3.7.5
numpy==1.24.4
packaging==24.2
pandas==2.0.3
pillow==10.4.0
prompt-toolkit==3.0.48
psycopg2-binary==2.9.10
pydantic==2.10.1
pydantic-core==2.27.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
redis==5.2.0
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.5.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
werkzeug==3.0.6
zipp==3.20.2
